+++ getdiskspaceosx/getdiskspaceosx/checkdiskspace.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+extern "C" {
+    bool getAvailableDiskSpace(const char* t_path, long* out_freeBytes);
+}