+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.3.0]
+### Added
+- getdiskspaceosx a native library for determining the amount of free space available on Mac OSX.
+- Status descriptions for updating app patcher state
+- Add configuration option to switch whether to use diffs or not
+
+### Changed
+- Separate version integrity check progress from update progress
+
 ## [3.2.0]
 ### Added
 - Using geolocation to find closest http server available